+++ client/src/scss/forms/_newMessageForm.scss
@@ -1,0 +1,11 @@
+.NewMessageForm {
+  height: 30px;
+  display: flex;
+  .textBox {
+    flex: 5;
+    padding-left: 5px;
+  }
+  .submitBtn {
+    flex: 1;
+  }
+}
