+++ client/src/scss/index.scss
@@ -7,6 +7,7 @@
 @import "./messagesViewport";
 // Forms styling
 @import "./forms/createAccountForm";
+@import "./forms/newMessageForm";
 
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;700&family=Roboto+Mono:wght@300;400;700&display=swap");
